+++ CHANGELOG.rst
@@ -2,9 +2,8 @@
 CHANGELOG
 =========
 
-
-1.18.2.dev
-==========
+1.18.2
+======
 
 * enhancement: Include SageMaker Notebook Instance version number in boto3 user agent, if available.
 * feature: Support for updating existing endpoint